+++ CHANGELOG.md
@@ -23,6 +23,7 @@
     - Renamed Quadratic cost function to Least Squares
     - Added option to stratify dataset in Hold Out and K Fold
     - Added Monte Carlo cross validator
+    - Implemented noise as layer instead of activation function
 
 - 0.0.4-alpha
     - Added Dropout hidden layer